+++ funky/app.py
@@ -5,12 +5,17 @@ containing directory.  In other words, use `python funky`.
 """
 
 import argparse
+from importlib.resources import read_text
+from pathlib import Path
 import sys
 from typing import List, Optional, Sequence, Union
 
 import funky
 from funky import commands, errors
 from funky.utils import log
+
+
+_SUPPORTED_SHELLS = ["bash", "zsh"]
 
 
 def main(argv: Sequence[str] = None) -> int:
@@ -26,6 +31,12 @@ def main(argv: Sequence[str] = None) -> int:
         log.logger.debug("Starting funky.")
         log.logger.vdebug("argv = {}".format(argv))  # type: ignore
         log.logger.vdebug("Command-line Arguments: {}".format(args))  # type: ignore
+
+        if args.init:
+            return run_init(args.init)
+
+        if args.setup_shell:
+            return run_setup_shell(args.setup_shell)
 
         _CmdAction.command(args)
     except errors.ArgumentError as e:
@@ -43,6 +54,39 @@ def main(argv: Sequence[str] = None) -> int:
     return 0
 
 
+def run_init(shell: str) -> int:
+    del shell  # all supported shells use the same code ATM
+
+    funky_sh = read_text("scripts.shell", "funky.sh")
+    print(funky_sh, end="")
+    return 0
+
+
+def run_setup_shell(shell: str) -> int:
+    if shell == "zsh":
+        config_base = ".zshrc"
+    else:
+        assert shell == "bash"
+        config_base = ".bashrc"
+
+    config_path = Path.home() / config_base
+    funky_init = f"funky --init {shell}"
+    if not config_path.exists() or funky_init not in config_path.read_text():
+        with config_path.open("a") as f:
+            cmd = (
+                "\n# setup funky\n"
+                "command -v funky &>/dev/null &&"
+                f' eval "$({funky_init})"'
+            )
+            log.logger.info(
+                "Appending %d lines to your %s file...",
+                len(cmd.split("\n")),
+                config_base,
+            )
+            f.write(cmd)
+    return 0
+
+
 def _get_argparser(verbose: bool = False) -> argparse.ArgumentParser:
     """Get command-line arguments.
 
@@ -54,6 +98,22 @@ def _get_argparser(verbose: bool = False) -> argparse.ArgumentParser:
         argparse.ArgumentParser object.
     """
     parser = argparse.ArgumentParser(prog="funky", description=funky.__doc__)
+
+    init_group = parser.add_mutually_exclusive_group()
+    init_group.add_argument(
+        "--setup-shell",
+        choices=_SUPPORTED_SHELLS,
+        help="Ensure your shell is configured correctly.",
+    )
+    init_group.add_argument(
+        "--init",
+        choices=_SUPPORTED_SHELLS,
+        help=(
+            "Initialize your shell environments. This should be run from your"
+            " .bashrc / .zshrc file."
+        ),
+    )
+
     parser.add_argument(
         "-d", "--debug", action="store_true", help="Enable debug mode."
     )