+++ tests/scripts/shell/test_funky.sh
@@ -6,9 +6,9 @@ if ! command -v shunit2 &>/dev/null && ! [[ -f ./shunit2 ]]; then
 fi
 
 OLD_HOME="${HOME}"; export HOME=/tmp/home
-export FUNKY_CMD="python ${PWD}/funky"
+export FUNKY_CMD="python3 -m funky"
 
-source ./scripts/shell/funky.sh 2> /dev/null
+eval "$(${FUNKY_CMD} --init bash)"
 
 oneTimeSetUp() {
     mkdir -p "${HOME}"